% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adjusted_r_squared}
\alias{adjusted_r_squared}
\title{Calculate adjusted r-squared}
\usage{
adjusted_r_squared(rsquared, n, p, method = c("ezekiel", "smith",
  "olkinpratt", "pratt", "wherry1", "wherry2"))
}
\arguments{
\item{rsquared}{the value of r-squared [0 to 1.0]}

\item{n}{sample size}

\item{p}{number of predictors}

\item{method}{character string indicating the method for calculating r-squared.
       One of \code{"ezekiel"} (default)
\code{"olkinpratt"}, \code{"pratt"},  \code{"wherry1"}, and \code{"wherry2"}}
}
\description{
Calculates adjusted r-squared using various methods.
}
\details{
The \code{"ezekiel"}  formula is typically used in statistics packages
(e.g., \code{adj.r.squared} in \code{\link{summary.lm}})
One recommendation is to use \code{"olkinpratt"} when the predictors are assumed to be random
and \code{"ezekiel"} when predictors are assumed to be fixed.
However, in practice differences between \code{"olkinpratt"} and \code{"ezekiel"} can be trivial.
All included formulas are designed to estimate population variance explained, \eqn{\rho^2}, rather than
sample variance explained, \eqn{R^2}.
}
\examples{
# This first example obtains argument values from sample data.
data(facets_data); data(facets_meta)
rs <- summary(regression('swl', ivs=facets_meta$ipip_factors,
                         data=facets_data))$r.squared
n <- nrow(facets_data)
p <- length(facets_meta$ipip_factors)

# These are then passed to the functions using different
#     methods.
adjusted_r_squared(rs, n, p, method='ezekiel')
adjusted_r_squared(rs, n, p, method='olkinpratt')
adjusted_r_squared(rs, n, p, method='pratt')
adjusted_r_squared(rs, n, p, method='wherry1')
adjusted_r_squared(rs, n, p, method='wherry2')

# Alternatively, the values of rsquared, p, and n
#     can be supplied directly.
adjusted_r_squared(.2, 100, 1)
adjusted_r_squared(.2, 100, 5)
adjusted_r_squared(.2, 100, 10)
}
\references{
Raju, N. S., Bilgic, R., Edwards, J. E., & Fleer, P. F. (1997).
Methodology review: Estimation of population validity and cross-validity, and the
use of equal weights in prediction. \emph{Applied Psychological Measurement, 21}(4), 291-305.

Yin, P., & Fan, X. (2001). Estimating R2 shrinkage in multiple regression:
A comparison of different analytical methods. \emph{Journal of Experimental Education, 69}(2), 203-224.

See also: \url{http://stats.stackexchange.com/a/63766/183}
}

