% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{facets_data}
\alias{facets_data}
\title{Data file with personality facets, personality factors, and well being outcomes}
\format{337 participants and 45 variables. Facets appear below with the naming
convention of "ipip_" then a letter corresponding to the factor that the facet belongs to
 (n=neuroticism, e=extraversion, o=openness, a=agreeableness, c=conscientiousness),
 then the facet name
\itemize{
\item  idhash. Identifier
\item  ipip_n_anxiety
\item  ipip_n_anger.
\item  ipip_n_depression
\item  ipip_n_self_consciousness
\item  ipip_n_immoderation
\item  ipip_n_vulnerability
\item  ipip_e_friendliness
\item  ipip_e_gregariousness
\item  ipip_e_assertiveness
\item  ipip_e_activity_level
\item  ipip_e_excitement_seeking
\item  ipip_e_cheerfulness
\item  ipip_o_imagination
\item  ipip_o_artistic_interests
\item  ipip_o_emotionality
\item  ipip_o_adventurousness
\item  ipip_o_intellect
\item  ipip_o_liberalism
\item  ipip_a_trust
\item  ipip_a_morality
\item  ipip_a_altruism
\item  ipip_a_cooperation
\item  ipip_a_modesty
\item  ipip_a_sympathy
\item  ipip_c_self_efficacy
\item  ipip_c_orderliness
\item  ipip_c_dutifulness
\item  ipip_c_achievement_striving
\item  ipip_c_self_discipline
\item  ipip_c_cautiousness
\item  ipip_neuroticism. Neuroticism factor
\item  ipip_extraversion. Extraversion factor
\item  ipip_openness. Openness factor
\item  ipip_agreeableness. Agreeablesness factor
\item  ipip_conscientiousness. Conscientiousness factor
\item  swl. Diener's Satisfaction with life scale
\item  panas_pa. PANAS positive affect
\item  panas_na. PANAS negative affect
\item  pwb_prelwo. Ryff's PWB - Positive relations with others
\item  pwb_autonomy. Ryff's PWB - Autonomy
\item  pwb_emastery. Ryff's PWB - Environemental mastery
\item  pwb_pgrowth. Ryff's PWB - Personal growth
\item  pwb_plife. Ryff's PWB - Purpose in life
\item  pwb_selfaccept. Ryff's PWB - Self-acceptance
}}
\usage{
data(facets_data)
}
\description{
The dataset includes 337 participants. The personality scales are based on the IPIP
scales modelled on the NEO-PI. Thus, there are scales for the Big 5 along with six facets for each
of the factors. There are also several well-being variablesli
}
\keyword{data}

