% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{facets_meta}
\alias{facets_meta}
\title{Meta data for the facets_data data frame}
\format{A list of character vectors corresponding to names in the facets_data data frame.
\itemize{
\item ipip_facets. A characeter vector of the 30 personality facet variable names.
\item ipip_factors. A character vector of the 5 personality factor variable names.
\item swb. A character vector the 3 subjective well-being variable names.
\item pwb. A character vector of the 6 psychological well-being variable names.
\item allscales. A character vector of all facet, factor, swb, and pwb scale variable names.
}}
\usage{
data(facets_meta)
}
\description{
Writing out the full list of facet variable names can be tedious.
Thus, it is often useful to store frequently used sets of variables
as elements in a named list.
This object stores these lists of names as elements in a list so that
they can be easily used in relevant functions such as \code{\link{regression}}
and \code{\link{bootstrap_r_squared_change}}.
}
\examples{
## Load the data from the package:
data(facets_meta)

## Print all the vectors of names:
facets_meta

## Print all facet then factor names:
facets_meta$ipip_facets
facets_meta$ipip_factors

## These lists can then be used in relevant functions:
data(facets_data)
summary(regression('swl', facets_meta$ipip_facets, data=facets_data))
}
\keyword{data}

