% Generated by roxygen2 (4.0.1): do not edit by hand
\name{facets_semi_partial_r_table}
\alias{facets_semi_partial_r_table}
\title{Semi-partial correlations between facets and criterion}
\usage{
facets_semi_partial_r_table(dv, facets, factors, data, return_pvalue = FALSE)
}
\arguments{
\item{dv}{character string of the criterion variable name in \code{data}}

\item{facets}{character vector of the facet variable names in \code{data}}

\item{factors}{character string of the factor variable names in \code{data}.
This string should be the same length as the facets list.
Thus, factor names should repeat and correspond to the factor for the facet
in the corresponding position in facets.}

\item{data}{data.frame}

\item{return_pvalue}{logical indicating whether to return p-values associated with each correlation}
}
\value{
By default a data frame of correlations is return. If \code{return_pvalue} is true then a list
of two data frames is returned one with the correlations and the other with the pvalues.
\enumerate{
\item \code{r_zero_order}: zero order corelation
\item \code{sr_focal_factor}: semi-partial correlation controlling only for focal factor
\item \code{sr_all_factors}: semi-partial correlation controlling for all factors
\item \code{sr_other_facets}: semi-partial correlation controlling for all other facets
}
}
\description{
Produce a data.frame of correlations between facets and criterion.
Rows represent each facet and the four columns represent the zero-order correlation between
facet and criterion followed by various semi-partial correlations where the facet is adjusted
controlling for the focal factor, all factors, and all other facets respectively
 as described under the "Value" section.
}
\details{
The table is useful for examining the degree to which facets provide a unique
contribution to predicting a criterion. In particular, facets controlling for factors is particularly
relevant to discussion regarding the incrmental value of facets.
}
\examples{
## Load data and meta data:
data(facets_data); data(facets_meta)

## Here we see how the facets and factors should be organised:
cbind(facets=facets_meta$ipip_facets,
       factors=rep(facets_meta$ipip_factors, each=6))

facets_semi_partial_r_table('swl',
    facets_meta$ipip_facets,
    rep(facets_meta$ipip_factors, each=6),
    facets_data)

## Return p-values
x <- facets_semi_partial_r_table('swl',
                                 facets_meta$ipip_facets,
                                rep(facets_meta$ipip_factors, each=6),
                               facets_data, return_pvalue=TRUE)

alpha <- .001
cbind(x$p[,1:2], ifelse(x$p[,3:6]<alpha, "*", ""))
}

