% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.bootstrap_r_squared_change}
\alias{print.bootstrap_r_squared_change}
\title{Print output for bootstrap_r_squared_change object}
\usage{
\method{print}{bootstrap_r_squared_change}(x, digits = 3, ...)
}
\arguments{
\item{x}{object of class bootstrap_r_squared_change}

\item{digits}{positive intenger: number of decimal points to display.
Numbers are passed to \code{round(x, digits)}}

\item{...}{further arguments passed to or from other methods (not currently used)}
}
\description{
Print output including descriptive statistics and bootstrap results.
}
\examples{
## Load data and meta data:
data(facets_data); data(facets_meta)

## Using 50 iterations is too few, but is used here to
## make the example run quickly.

##  Save object
fit <- bootstrap_r_squared_change(facets_data,
             facets_meta$swb[1],
             facets_meta$ipip_factors,
             facets_meta$ipip_facets,
             iterations=50)
## print object
print(fit)

## Alternatively, the object fit object can be explored directly:

## Show density plot of bootstrapped estimates of rho-squared change.
plot(density(fit$theta_hats),
     xlab="rho-squared change",
     main="Density of bootstrap estimates")

## Extract the confidence interval values
fit$ci_values
}

