\docType{package}
\name{personalityfacets-package}
\alias{personalityfacets}
\alias{personalityfacets-package}
\title{A package for bootstrapping differences between mean correlations}
\description{
  This is currently just a test package that I'm using to
  learn about package development
}
\author{
  Jeromy Anglim
}

