\docType{package}
\name{personalityfacets-package}
\alias{personalityfacets}
\alias{personalityfacets-package}
\title{A package for obtaining bootstrap estimates of confidence intervals on population r-square change}
\description{
  Package performs double adjuste rsquare bootstrap on
  r-square change.
}
\author{
  Jeromy Anglim
}

